py/requirements.txt
playwright
beautifulsoup4
lxml
requests
supabase
python-dotenv
